- Ensured that floating point constraits are rounded in right direction - Improved Z3 session management

+++ SNNT/src/Parsing/Parsing.jl
@@ -253,7 +253,7 @@ function parse_factor(tokenmanager :: TokenManager)
 		@debug "Parsing number"
 		current_token = next(tokenmanager)
 		#TODO(steuber): FLOAT INCORRECTNESS
-		return TermNumber(parse(Float64, untokenize(current_token)))
+		return TermNumber(parse(BigFloat, untokenize(current_token)))
 	elseif Tokens.kind(current_token) == Tokens.IDENTIFIER
 		@debug "Parsing variable"
 		current_token = next(tokenmanager)
